+++ src/SW_Control.c
@@ -1738,7 +1738,6 @@ void SW_CTL_read_inputs_from_disk(
 
     SW_LYR_read(
         &sw->RunIn.SoilRunIn,
-        &sw->SiteSim.n_evap_lyrs,
         &sw->RunIn.SiteRunIn.n_layers,
         SW_PathInputs->txtInFiles,
         LogInfo