+++ src/SW_Main_lib.c
@@ -546,7 +546,6 @@ void sw_setup_prog_data(
     // initialize output
     SW_OUT_setup_output(
         SW_Domain->nMaxSoilLayers,
-        SW_Domain->nMaxEvapLayers,
         sw_template->VegEstabIn->count,
         totNSites,
         &sw_template->VegEstabIn->parms,