+++ src/SW_VegProd.c
@@ -1265,7 +1265,6 @@ void SW_VPD_read(
 
     const char *const lineErrStrings[] = {
         "vegetation type components",
-        "vegetation type components",
         "albedo values",
         "canopy xinflec",
         "canopy yinflec",
@@ -1284,11 +1283,14 @@ void SW_VPD_read(
         "shade yinflec",
         "shade range",
         "shade slope",
+        "rooting profile: shape parameter 'a'",
+        "rooting profile: shape parameter 'b'",
+        "rooting profile: depth of rooting zone",
         "hydraulic redistribution: flag",
         "hydraulic redistribution: maxCondroot",
         "hydraulic redistribution: swpMatric50",
         "hydraulic redistribution: shapeCond",
-        "critical soil water potentials: flag",
+        "critical soil water potentials",
         "CO2 Biomass Coefficient 1",
         "CO2 Biomass Coefficient 2",
         "CO2 WUE Coefficient 1",
@@ -1304,7 +1306,7 @@ void SW_VPD_read(
     int lineno = 0;
     int index;
     // last case line number before monthly biomass densities
-    const int line_help = 32;
+    const int line_help = 35;
     double help_veg[NVEGTYPES];
     double help_bareGround = 0.;
     double litt;
@@ -1316,7 +1318,6 @@ void SW_VPD_read(
     char inbuf[MAX_FILENAMESIZE];
     char vegStrs[NVEGTYPES][20] = {{'\0'}};
     char bareGroundStr[20] = {'\0'};
-    char *startOfErrMsg;
     const int numMonthVals = 4;
     int expectedNumInVals;
 
@@ -1329,14 +1330,12 @@ void SW_VPD_read(
     while (GetALine(f, inbuf, MAX_FILENAMESIZE)) {
         lineno++;
 
-        startOfErrMsg = (lineno >= 27) ? (char *) "Not enough arguments" :
-                                         (char *) "Invalid record in";
-
         if (lineno <= line_help) {
-            if ((lineno >= 1 && lineno <= 3) || lineno == 31 || lineno == 32) {
+            if ((lineno >= 1 && lineno <= 3) || lineno == 34 || lineno == 35) {
 
                 x = sscanf(inbuf, "%19s", vegStrs[0]);
                 expectedNumInVals = 1;
+
             } else {
                 // Inputs must match order of veg types 0..NVEGTYPES
                 x = sscanf(
@@ -1373,10 +1372,13 @@ void SW_VPD_read(
                 LogError(
                     LogInfo,
                     LOGERROR,
-                    "%s %s in %s\n",
-                    startOfErrMsg,
-                    lineErrStrings[lineno - 1],
-                    MyFileName
+                    "Invalid input %s in '%s': "
+                    "found n = %d values but expected %d on input line %d",
+                    lineErrStrings[lineno],
+                    MyFileName,
+                    x,
+                    expectedNumInVals,
+                    lineno
                 );
                 goto closeFile;
             }
@@ -1560,34 +1562,53 @@ void SW_VPD_read(
                 }
                 break;
 
-            /* Hydraulic redistribution */
+            /* Rooting profile parameters */
             case 22:
+                ForEachVegType(k) {
+                    SW_VegProdIn->veg.rootProfileParam[k][0] = help_veg[k];
+                }
+                break;
+
+            case 23:
+                ForEachVegType(k) {
+                    SW_VegProdIn->veg.rootProfileParam[k][1] = help_veg[k];
+                }
+                break;
+
+            case 24:
+                ForEachVegType(k) {
+                    SW_VegProdIn->veg.rootProfileParam[k][2] = help_veg[k];
+                }
+                break;
+
+            /* Hydraulic redistribution */
+            case 25:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.flagHydraulicRedistribution[k] =
                         (Bool) EQ(help_veg[k], 1.);
                 }
                 break;
 
-            case 23:
+            case 26:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.maxCondroot[k] = help_veg[k];
                 }
                 break;
 
-            case 24:
+            case 27:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.swpMatric50[k] = help_veg[k];
                 }
                 break;
 
-            case 25:
+            case 28:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.shapeCond[k] = help_veg[k];
                 }
                 break;
 
             /* Critical soil water potential */
-            case 26:
+            case 29:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.SWPcrit[k] = -10. * help_veg[k];
                     // for use with get_swa for properly partitioning swa
@@ -1598,14 +1619,14 @@ void SW_VPD_read(
 
             /* CO2 Biomass Power Equation */
             // Coefficient 1
-            case 27:
+            case 30:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.co2_bio_coeff1[k] = help_veg[k];
                 }
                 break;
 
             // Coefficient 2
-            case 28:
+            case 31:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.co2_bio_coeff2[k] = help_veg[k];
                 }
@@ -1613,21 +1634,21 @@ void SW_VPD_read(
 
             /* CO2 WUE Power Equation */
             // Coefficient 1
-            case 29:
+            case 32:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.co2_wue_coeff1[k] = help_veg[k];
                 }
                 break;
 
             // Coefficient 2
-            case 30:
+            case 33:
                 ForEachVegType(k) {
                     SW_VegProdIn->veg.co2_wue_coeff2[k] = help_veg[k];
                 }
                 break;
 
             /* Spatial reference of biomass inputs */
-            case 31:
+            case 34:
                 SW_VegProdIn->isBiomAsIf100Cover =
                     sw_strtoi(vegStrs[0], MyFileName, LogInfo) ? swTRUE :
                                                                  swFALSE;
@@ -1637,7 +1658,7 @@ void SW_VPD_read(
                 break;
 
             /* Calendar year corresponding to vegetation inputs */
-            case 32:
+            case 35:
                 SW_VegProdIn->vegYear =
                     (TimeInt) sw_strtoi(vegStrs[0], MyFileName, LogInfo);
                 if (LogInfo->stopRun) {
@@ -1652,11 +1673,8 @@ void SW_VPD_read(
         } else {
             /* Mean monthly vegetation inputs */
 
-            if (lineno == line_help + 1 || lineno == line_help + 1 + 12 ||
-                lineno == line_help + 1 + 12 * 2 ||
-                lineno == line_help + 1 + 12 * 3 ||
-                lineno == line_help + 1 + 12 * 4 ||
-                lineno == line_help + 1 + 12 * 5) {
+            if (lineno >= line_help + 1 &&
+                ((lineno - (line_help + 1)) % 12) == 0) {
                 mon = Jan;
             }