+++ NEWS.md
@@ -1,4 +1,33 @@
 # NEWS
+
+# SOILWAT2 v8.4.0-devel
+* Simulation output remains the same as the previous version.
+  However, output of the example differs slightly from previous example output
+  because of the following changes: updated potential evaporation coefficients;
+  updated rooting profiles.
+
+* Rooting profiles can now be estimated with an equation and parameters
+  for each vegetation type; soil impermeability represents (semi-)restrictive
+  soil layers that reduce rooting profiles (#459; @dschlaep).
+
+* Potential evaporation coefficients can now be estimated from soil properties
+  (#409; @dschlaep).
+
+## Changes to inputs
+* New input via `"siteparam.in"` to select the input option for potential
+  evaporation coefficients: either provided as inputs by `"soils.in"` or
+  estimated from soil properties.
+* New input via `"siteparam.in"` to select the input option for
+  rooting profiles: either provided as inputs by `"soils.in"` or
+  estimated with an equation and parameters for a vegetation type.
+* New inputs via `"veg.in"` to provide values for the three parameters of the
+  rooting profile equation for each vegetation type.
+
+## Changes to outputs
+* All modes (including txt-based) now output values of soil evaporation
+  at each soil layer.
+
+
 # SOILWAT2 v8.3.0
 * Simulation output remains the same as the previous version.
   However, output of the new plant functional types are renamed and include